Add animation toolbar for motion layout and animated-vector
When opening a motion layout file in split editor, we should display an
animation toolbar (guarded by the NELE_MOTION_LAYOUT_ANIMATIONS flag) on
the bottom of the visual portion of the editor. The animation toolbar
should have a seek bar so we can jump to specific portions of the
animation.

Similarly, when inspecting vector drawables, we should display an animation
bar as well. The toolbar is guarded by the NELE_ANIMATIONS_PREVIEW flag and
does not have a seek bar.

This reapplies change I547b2842dc50f6807b46816c8aa5d67fa6a744c9, which
made GuiTestRuleTest#makeSimpleApplication fail, since it adds a motion
layout XML file to the SimpleApplication test project. The problem is we
actually try to make this project, which does not supportmotion/constraint
layout. In this CL, we moved motion_layout.xml to SimpleAndroidXApplication,
which is enough to make the tests pass.

Bug: 139007163
Test: AnimationToolbarTest
Change-Id: Ib1e57fe71efd862a6ee67469cc444b6c7c79f36c