Warn when proxy authentication requires a password
... in the gradle.properties file.

When users setup a proxy for Gradle, we show a message that they can
setup proxy passwords manually. If proxy authentication is used but no
password is defined it is very likelly that there will be connection
errors from Gradle.

This CL shows a warning when proxy authentication is enabled and no
password exists already on the gradle.properties file. The warning comes
with an hyperlink to open the global grdle.properties file.

Bug: 135102054
Test: Added testApplyHttpProxySettingsWithAuthenticationWithPassword and
adapted existing ProxySettingsDialogTest tests.
Change-Id: I322b2929c44ef8b1ca1dda509455026b9ef31506