Show Proxy dialog only if no host and port
... are included in gradle.properties when using proxy auto
configuration.

When using PAC in the IDE we cannot tell if the settings passed to
gradle are or not the same since gradle only supports manual settings.

Instead of always showing the proxy settings when PAC is used, the
dialog is only shown when either the host or the port are not defined.

Bug: 135102054
Test: testShowProxySettingDialogPAC
Change-Id: I8b3c7aee971cd7ed1a830f1d4bcbdd5962668034