Refactor change detection on Compose Preview
This replaces the change detection so it does not use a
PsiChangeListener but a DocumentListener. Document listeners are
specific to the file so they receive less traffic.
We now save smart pointers to the two relevant areas for the
PreviewElement, the element definition (e.g. @Preview()) and the preview
body (the actual contents): This allows to detect where the changes are
being done.

Bug: N/A
Test: Updated tests
Change-Id: Icc3f32080c03aa2ef673428b792d1aaeb41f87d5