Do not require an AndroidFacet to accept design files
When creating a new project using the new project wizard, we open a
layout file when indexing is done, but before sync happens. Since we
require an AndroidFacet to determine if the file should be open by the
split editor, we're gonna reject the file as the facet will only be
available after sync. As a result, we open the layout file using the
default text editor, and design tools will only be available if the user
close and reopen the file.

We don't actually need an AndroidFacet to figure out if a file can be
open by the split editor, since we only need the PSI structure (available
after indexing) to detect the file types we support. Simply removing the
requirement of AndroidFacet from DesignerEditorProvider#accept it not
enough, as we would then fail when adding a new module (see b/111554518).

The fix for b/111554518 got rid of the exception thrown when adding a new
module, but introduced a similar problem to creating a new project: when
we add a new module, DesignerEditorProvider#accept will reject the files
newly added, since the AndroidFacet won't be available by the time they're
opened. Again, the user would need to close and reopen the file to use the
design tools.

In order to fix the issue for both new projects and modules, we do the
following in this CL:
  * Don't require an AndroidFacet in DesignerEditorProvider#accept
  * When we don't find an AndroidFacet when initiating the NlModel,
    listen to the sync that will hapen *after* creating the module and
    opening the files. This sync is responsible for triggering the
    AndroidFacet creation for the module.
  * When we remove a module while keeping a resource file belonging to it
    open, remove the model from the surface. That will prevent users from
    interacting with a model from a module already removed from the project,
    which would cause exceptions to be thrown and the UI to be inconsistent
    and/or broken.

Fixes: 139781107
Bug: 111554518
Test: updated unit tests. Manual test steps are in the bugs above.
Change-Id: I08e2ae268f9fcce1aa0ce80f6d7aafebb1d1d214
(cherry picked from commit 4c2e65182f2930d007ecb4b0c107d6009af98ecb)