[Gradle Sync] Ensure GradleFiles listener is added on open
This change ensures that the listener for GradleFiles is correctly added
to the psi trees when a project has been opened. We need to do this
because the GradleFiles object is constructed after the
FileEditorManagers events has been fired. This means we miss the event.

This change triggered this check manually so that we correctly add the
listener.

Bug: N/A
Test: manually tested
Change-Id: I79dda97a433261aa1808116410fbf34e22d6cdb9