Do not cache SceneView statically
SceneView may have the references to all the layers includes
ScreenViewLayer, which we used to store the preview image
rendered by Layoutlib. The image needs couple of memory and
we want it to be released when JVM needs some memory.

We used a weak map with SoftReference to cache the
SceneContext for a SceneView. But the cached SceneContext
has a reference to SceneView and SoftReference only be
released when JVM almost runs out of memory. This results
we run out of memory first then release SceneView later.

But it is a bad expereinces for users since Studio becomes slow
when JVM runs out of memory. Although it will be recovered after
JVM releases the SoftReferences, but there is no reason to make
user wait and suspect the Studio is hanging.

Also the doesn't reuse SceneView and it is always used with
SceneView, it doesn't need to be cached.

Bug: N/A
Test: covered by exsiting test
Change-Id: I307c33dc28cae3f64941c2934e96cb6bf70e98db