Improve LightBindingClass cache consistency
Previously, we cached LightBindingClass lists against a PSI file, with
the intent that those classes would get cleaned up when the file was
deleted.

However, in practice, this regularly resulted in the cache getting
rebuilt, which meant we kept generating new versions of existing
LightBindingClass instances. Unfortunately, this ended up being
problematic, as the different instances were not considered equivalent
and did not work with existing scope enlarging code.

Bug: N/A
Test: Covered by existing
Change-Id: If3f95ae17476db2b0996279cc45632d6dceac035