Categorize launch failures for unknown, missing activities
If an ActivityLaunchTask fails for any reason, these errors currently
turn up in our logs as RunEvents with empty deploy_failure_ids. We
recently saw a spike in deploy failures due to a regression in
DefaultActivityLocator that was difficult to pin down because these
failures didn't supply a descriptive deploy_failure_id.

This CL makes our deploy failure metrics more meaningful by setting
the failure id to "UNABLE_TO_DETERMINE_LAUNCH_ACTIVITY" if the
ActivityLocator returns null for the activity name, or setting it
to "ACTIVITY_DOES_NOT_EXIST" if for some reason the ActivityLocator
chooses an activity that isn't present in the installed APK. If the
launching the choosen activity fails for some other reason, we set
the failure id to "UNKNOWN_ACTIVITY_LAUNCH_TASK_ERROR" so that we
can at least isolate the failure to ActivityLaunchTask.

Test: ActivityLaunchTaskTest
Bug: 139421849
Change-Id: I1488453bce5adb075dd205ed0e8bc3c2573b6aab