Update JDK warning visibility more often
When the user has JAVA_HOME selected as JDK location and then changes it
to embedded we show a warning label with a link to prevent that this
could cause multiple Gradle daemons. This should also happen when
selecting STUDIO_GRADLE_JDK or when the user types the path and tries to
apply changes.

This CL also hides the warning when the JAVA_HOME value is not valid.

Bug: 141110179
Test: Manual inspection
Change-Id: Iab057aede2369e65931b2261b88cb618c3fe1958