Add explict WorkBench parent disposable
Make a parent Disposable part of the WorkBench constructor. WorkBench
needs to have one to correctly dispose but the contract did not enforce
this in any way so, in a few places, we had potential leaks for
WorkBenches that would never be registered to a parent.
In most cases we just want the WorkBench to be attached to the parent
editor.

Fixes: 140160789
Test: Existing tests pass
Change-Id: I1fa5688c65e1ef7107a529145db9a2ff6114bb63