Add tootltip and link to JDK description
... to give users more context on what this setting does.

Also made some changes to simplify how the notifications are shown when
JAVA_HOME is not being used as JDK location.

Use SelectJdkFromFileSystemHyperlink instead of UseJavaHomeAsJdkHyperlink
to give users more control.

This also removes a link and a message from the IdeSdksConfigurable form
that is now redundant after adding a ContextHelpLabel.

Also made similar changes to JdkSetupStep to have a closer look and
feel.

Bug: 141110179
Test: Manually confirmed changes in messages and UI.
Change-Id: Id26e800ffe8be7413d314290df9090a21b78f4bf