Expose DefaultActivityLocator.computeDefaultActivity
DefaultActivityLocator.computeDefaultActivity is a useful
utility function that should be exposed.  In this case, the
blaze plugin uses it to calculate the default activity, instead
of re-implementing this same logic.

computeDefaultActivity currently takes an option Device parameter.
This forces the method to be annotated with @Slow, even though that
only happens when a device is given.  computeDefaultActivity has now
been split into two methods, computeDefaultActivity and
computeDefaultActivityWithDevicePreference.  The first is fast and can
be called from UI thead, while the second is slower and should not.

@Slow and @WorkerThread annotations have been added to methods
that may hang waiting on an IDevice to respond.

Bug: 133244915
Change-Id: If6511ffafc4b5a3f20dec3572efa9bdec8c67533
Test: Added DefaultActivityLocatorUnitTest. Other existing tests should still pass.