Process spaces and line numbers in console links
... from build output.

We currently have GenericFileFilter.kt to add hyperlinks for files in
the build output but we are missing two cases:

1.- When the path comes with line and column numbers inside a parenthesis,
like "/path/to/file: (line, column):"
2.- Spaces in the path, like "/path/to/file/with space.kt"

For the second case, the generic filter now looks for the " (" pattern
after a colon is found and in that case it tries to process the numbers.

For the second case, when processing a path and a white space is found,
now it tries to create the link to an existent file and if it does not
find one the continues with the path as if the space were part of it.

Bug: 138131992
Bug: 136242040
Test: Added and adapted tests in GenericFileFilterTest.kt
Change-Id: I196fbb4a31d7a63020fad01e38ce5e838216be55