Restore the selected key frame if possible.
This will help the properties panel to update when changing
attributes on a key frame. It will only work for attributes that
are not key attributes. Another solution is needed when a key
attribute is updated. Perhaps move the selection to the transition.
That will be done in a followup CL.

Bug: 118388686
Test: TODO
Change-Id: I06b72d22c4ab7fcb754c32b75e5e87202b1285ec