[Res Explorer] Resolve getConfiguration() uses in the EDT
This fixes the memory leak in I2b83871f, which was reverted in
I7547d28a.

In some instances, particularly with the New Resource Picker, the
Resource Explorer would try to get a configuration (@Slow method) under
the EDT.

These changes make it so the ResourceExplorerListViewModelImpl doesn't
provide any resource data to ResourceExplorerListView until Configuration
and ResourceResolver are initialized, should avoid blocking the
UI-thread.

Refactored the UI that would contain the list of resources into
ResourceExplorerListView and its view model
ResourceExplorerListViewModel, these will be created every time the
AndroidFacet is changed, and they won't be created until
ResourceResolver is initialized.

ResourceExplorerView will consist on just the resource types TabbedPane
and the bottom actions (grid/list mode, zoom in/out).

ResourceExplorerViewModel will now have a method that will create
ResourceExplorerListViewModel asynchronously to make sure it's
initialized with a Configuration and a non-null ResourceResolver, this
ResourceResolver should now be the only one used accross the
ResourceExplorer, for both rendering and displaying resource data.

There shouldn't be any actual change in behavior or features from the
user's pov.

Bug: 140234608
Bug: 140235470
Test: added and updated some unit tests, ui-tests with no complains in aspects agent log
Change-Id: I5992186cc9ce88d555502f28df4f8b1c081a816c