Remove LaunchTask#perform API
The LaunchTask interface defines a run() method for executing the task
and obtaining a LaunchResult. The default implementation of run()
delegates to LaunchTask#perform for launching the task and wraps the
boolean result with a LaunchResult. In this setup, callers should always
use run() instead of directly calling perform(), but we can't enforce
this by making the method protected because LaunchTask is an interface.

This CL resolves the issue by removing the perform() API altogether and
and inlining the default implementation of run() in all the implementing
classes.

Test: existing (no behavioral changes)
Bug: 140403914
Change-Id: I9b04e0260477b4a123c0a9f8902561caa08d6a52