[ui-test] Fix SideConstraintHandlingTest
Test consistently fails, caused by trying to create constraints by
clicking and dragging but getting the constraint pop-up menu instead.
This is caused from wrongly predicting the position of the anchor due to
having the scene view in a very restricted space. This fixes the test by
closing the project tree file panel, allowing for more space to the
LayoutEditor and getting the constraint anchors where they are expected.

More of a workaround than a fix. Maybe for a proper fix we could try to
calculate the position of the anchor based on the dimension of the scene
view.

Fixes: 124383902
Test: passes remotely
Change-Id: I9c4353ace9f0463c280b123ddfcd6a09161221e0