Fixed a bug concerning not null columns without default values
Only newly added columns which are not null and don't have a
default value request a complex update. Because of a badly places
if statement, we used to generate a complex update for any column
present in the final version of a table which was not null and without
a default value.

Test: SqlStatementsGeneratorTest
Bug: 136048932
Change-Id: I1741862be84773b77b88f40908c0a1ae05028d05