[PSD,.kts] support PSD kotlinscript resolution
The Project system performs reference resolution independent of
direct parsing, for example in resolving versions in
FakeArtifactElement.

The way it does this is perhaps unsound: it performs its own,
straightforward string manipulations of the syntax variable, rather
than going through the language-specific parser.  For Groovy this
works straightforwardly, because the implicit Dsl syntax and the
Groovy syntax are in one-to-one correspondence.  For Kotlinscript, we
need to perform the same intercession for these constructed references
as we do for references that we discover through the regular Dsl
parsing interface (using gradleNameFor())

Therefore, define a new GradleDslParser method, convertReferenceText,
to perform this conversion.  The default implementation, suitable for
the Groovy Dsl parser, is the identity function; the Kotlin parser can
use gradleNameFor(), though there is an impedance mismatch in that
gradleNameFor operates on Psi, whereas resolveReference operates on
(String) text.  We therefore somewhat wastefully convert the string
back to language-specific (Kotlin) Psi, in order to write the
equivalent in the Groovy-based syntax.

This has been manually tested by verifying that the PSD can correctly
display resolved information corresponding to the version of the
kotlin stdlib encoded as rootProject.extra["kotlin_version"]

Bug: 130224064
Test: existing, manual (see CL description above)
Change-Id: Id538ddab58f30f6790eff24cd2b1b5f9890cda3d