[Dsl,.kts] more extra property test cases
Now that we support qualified references to extra properties, and
more-or-less correctly handle the differences between Groovy's
"bubbling up" extra property resolution and Kotlin's lack of it,
implement and enable some more test cases for Kotlin.

The lack of bubbling up means that we need to fully qualify some extra
property lookups from subprojects, which in turn means that we need an
additional extraName method (including a container argument) to smooth
over the differences between Groovy and Kotlin syntax in our test
cases.

Bug: 130224064
Test: ported more ExtModelTest cases to Kotlin
Change-Id: Ic0caa85a4c141cb5c5edac89d0fc02c5d2290798