[Build Model] Remove use of GradleBuildFile
This change removes a usage of the old parser. Unfortunatly the new
parser is too slow to be used in this context. We instead switch to
first checking the models for the information required. If this is not
present then we revert back to a simple string search for the plugin ids
that should be applied. This should cover almost every case that the old
parser managed to cover. Plus cases where we can only find the type of
Project from the models.

Bug: 136801101
Test: tests will follow in a future CL
Change-Id: I242ba874e7b76d52644d5e110b36d6316e629d28