Remove warning message for WNA null scrolling
After exploring other ways of fixing the out-of-order events, it seems
like keeping the null check is still the simplest solution.

These conditions can cause the exception:
1. Open WNA such that the scrollbar is visible, close WNA, adjust window
height to be large enough to not need the scrollbar, then reopen WNA
from the side tab.
2. Open WNA such that scrollbar is visible, scroll to bottom, close WNA,
adjust window height so scrollbar would still be visible, then reopen
WNA from the side tab.

In both cases, the scroll handler is called before the metrics are
initialized because the reopen listener is wrapped in an invokeLater
(see ag/8446997). The other solution would be to wrap scrolledToBottom
in invokeLater too, which would be more correct, but metrics for
scrolling probably isn't important enough to warrant notifying the user
of possible crashes.

Bug: 140218425
Test: N/A
Change-Id: Ic856d7565ed95da077b9ae83e942c8c1dd5b62b6