Move to use new androidx package from ui-tooling
This change is needed for the preview to use the right packages from the
ui-tooling library. This will move the package names from the old
internal package to the androidx one used by ui-tooling.
Also add a notification to warn internal users of their project using
the old names.

Add units to width and height for the @Preview annotation.

Bug: 140879308
Test: Updates test
Change-Id: I0d4e03f6df4b2613721f2b44e8706bf27ca06307