Add "context" property to WorkBench
Add a context property to WorkBench, since it can be used in different
contexts. For example, in the split editor we might have a tool window
of a particular work bench being hidden in design mode but shown in split
mode. Adding the context allows us to restore the state corresponding to
the target context.

This reapplies change Ib2c28028c1c and fixes the UI tests that started
failing after that change.

Bug: 141176929
Test: AttachedToolWindowTest
Change-Id: If3495bb00f78576905c6a902bce0c227bcc82be7