Get method reference in the form of "var.method"
Before this CL, we only add references to methods in
the form of "var::method" and can not distinguish
them from method calls.
This CL adds method references for both "var::method"
and "var.method" and distinguish them from method
calls with an extra properrt "Kind" so that they can
be matched with their types by signature.
This CL also remove the case that a method can be
treated as a field with the same name. For example,
a field "something" may reference getSomething()
or isSomething but not something().

Fixes: 140592807
Test: DataBindingInspectionTest
Change-Id: I90ea0871af270ee8e8de04ffe4c2c270b4f85654