Ensure StudioFlag overrides are cleared between test cases
Any time a test case overrides a flag value, it should clear
the override at the end of the test to prevent unexpected behaviors
in other test cases. This can happen either in a tear-down method
or in a finally block when the entire test case runs inside the
corresponding try block (it isn't sufficient to just clear the
override at the end of the test outside of a try-finally because
there may be an exception or a failed assert).

This CL updates various tests in the codebase to follow this pattern.

Bug: 141274667
Test: these are the tests
Change-Id: Id697d960eebc3d85c70b13af63569967fa3e9ceb