[Build Model] Remove use of GradleBuildFile...
...from the CreateLibraryFromFilesAction. We also take this oppotunity
to improve this old action and fix some minor UX problems.

Firstly since it is possible for the new parser to be slow we show a
progress dialog for if this ends up taking a substantial amount of time,
this ensures the user isn't left in the dark about what is going on. For
most normal projects <a few hundred modules this dialog should never be
shown.

Secondly we fix an issue with the combo box drop down when chosing a
module to select. This change makes it so that only modules that are
android modules are shown and now correctly sets the initial selection
to be one of these.

Finally we fix 2 minor issues:
  1 - The same Jar was being added twice due to root detection also
  finding different OrderEntries for each scope. We now deduplicate
  paths before adding them to the build file.
  2 - The relative path was being incorrectly calculated.

Bug: 136801101
Test: Mannually tested
Change-Id: I6579ffd1ae64113b3d62a528e43e04288cf08a7d