UserCounterModel to create a new RangedContinuousSeries for each event specified by the user
Context:
Demonstrates how the UserCounterModel will be used. In the EventProfilerStage, a
UserCounterModel is made for each event the user is tracking. The UserCounterModel is added
to the track model, and in the CustomEventTrackRenderer the UserCounterModel is accessed
through the CustomEventTrackModel so that the line chart can be drawn for the specific event

Bug: 140522169
Test: UserCounterModelTest.kt
Change-Id: If5713fa02965559772d886121270320dbbd844b1