[Kotlin DSL] support signingConfig and add test coverage for Androidmodels.
This CL adds support to parsing and writing signingConfig elements, in
addition to the writer support for ProductFlavor elements. It
also adds test coverage for BuildTypes, and rewrite invalid KTS test
data files.

More test coverage for the writer for androidModels are coming in future
CLs. Tests for signingConfigs also will be added once we support
gradleNames from elements with localMethoName.

Bug: N/A.
Test: updated.
Change-Id: Icc6f7ea2ea3a5d97f25d473b866e62a7b0f48ecb