[Kotlin DSL] Add tests support to SigningConfig.
This CL adds test coverage to signingConfig dslElement. We need to make
sure we use the rawValue that corresponds to the build language type.
For example: in Kotlin we use localMethod names to refer to a signingConfig element and this methodName is sat, by default, to getByName(): assuming the user always gets a signing config element that exists.

Bug: N/A
Test: updated and data files added.
Change-Id: Id75d656bfc720f7011d56d9e2007e6fb8245409f