[Dsl] translate BuildType properties
Define groovyToModel and ktsToModel maps in BuildTypeDslElement, which
convert from source language expression of properties to internal
model names.

The properties in question correspond to model properties:

- debuggable
- embedMicroApp
- jniDebuggable
- minifyEnabled
- pseudoLocalesEnabled
- renderscriptDebuggable
- renderscriptOptimLevel
- shrinkResources
- testCoverageEnabled
- zipAlignEnabled

Rename the internal model string names to be distinct from both Groovy
and Kotlin names, as a debgging convenience to help keep track of
which world a particular string name comes from, external or
model.  (If everything has been done right, the model names are now
essentially arbitrary).

When adding a property to a BuildTypeDslElement, check if its external
name is an external name for one of our properties, and if so,
canonize the element to its model name before adding it.

The additional ExtModelTest tests that we do not do renaming in the
wrong context: an extra property "isDebuggable" should not be renamed,
even though it shares a name with a BuildType property.  It is not as
strong as ideal; it should also test the effect of an assignment to
the extra property, using

  ext.isDebuggable = false // Groovy
  isDebuggable = false // Kotlin

but at present assignment through a variable in Kotlin does not update
the extra property in the Dsl.

There are some cases in the Kotlinscript files that use and/or emit
invalid syntax, manifestations of b/141548181, fixed in subsequent
CLs in this chain.

Bug: 130224064
Bug: 141548181
Bug: 141842964
Test: strengthen existing BuildTypeModel tests (check file contents)
Test: enable more BuildTypeModel tests
Test: new ExtModelTest regarding property names
Change-Id: Ie7a06bd9b4ec24221315c4341b5c46a42c35b3b8