Remove log noise for invalid drawables
If an invalid drawable is passed, renderDrawable might throw an
exception. We can not fully control the input passed to this
rendering call since the user might be referencing an invalid
drawable so we are just less verbose about it. The user will not
see the preview next to the code when referencing invalid drawables.

Test: Manually checked the log while adding a new Button into the design
surface. This used to cause a warn log.
Fixes: 135525027

Change-Id: Iee28a9def9ef6c24548002a72a476847abd80ab2