Add PID as hint when parsing Atrace/Perfetto captures
With perfetto sometimes the process names are not dumped in the first
few seconds. As such getting a capture < 5 seconds it would ask the user
to select the process. Its not obvious what PID the user should select
but the profilers know this information. This changes passes the PID
in addition to the process name as a hint to the parser on which process
should be selected.

Test: Updated CpuCpatureStageTest
Bug: N/A

Change-Id: I1871ec7682c6c5e9eeb7d283f6ed236b86d141c8