Do not send configuration update event after refresh
A force refresh was sending a forced configuration udpated event. If the
configuration had not changed, it seems there is not need for that event
to happen.
When configuration changes it also forces the zoom to be reset so that
was causing weird zoom resets that were not expected by the user.

Fixes: 140878263
Test: Manully checked, existing tests pass
Change-Id: Iaa1852bd6f24ff5a09555b29db683882816d533e