Creating custom scene hierarchy
Results from Compose stack is not associated with the xml nor any
traditional layoutlib. To display the correct interactable preview, we
override the model updater and create a custom scene graph (using the
given bounds from compose, which right now is not exactly correct).

Bug: 141027052
Test: Able to click the preview and goto the source specified by the
compose stack (which right now is off due to inline functions)

Change-Id: I2516a309d0cca784f3d66f5a8f05755682940f4f