Migrate to language-specific modification trackers
This is in response to the deprecation of JAVA_STRUCTURE_MODIFICATION_COUNT
(see the linked bug).

Many of the cached values that relied on the Java structure modification
count did so because their values were expensive to compute. So,
we migrate to language-specific modification trackers in order to
keep performance reasonable while typing. In some cases this is a
temporary fix; the longer term fix is to make the values cheaper to
compute in the first place.

It's true that PsiModificationTrackerImpl#forLanguage is still marked
@ApiStatus.Experimental. However, we have helped JetBrains fix bugs in
it upstream (IDEA-219943), and the API now has good test coverage
in PsiModificationTrackerTest.JavaLanguageTrackerTest.

This CL also adds a new cache for RenderErrorContributor.getAllViews(),
thereby avoiding a class inheritor search on every keystroke when typing
out a tag name in a layout XML file. This was a major bottleneck
in some recent thread dumps from 3.5.

Test: PsiModificationTrackerTest and other existing tests
Bug: 138591235
Change-Id: Ia9f3b0110701019fd6bda5e133b4def407a1db1e
(cherry picked from commit 2547ab02bb2fa76960004287376380e4cca9f78f)