Fix problem with initial expansion size.
The UI was expanding nodes that should not be expanded by the
extent of the tree depth indent. It turns out that the initial
widths are computed before we can access the nodes through the path.
The expansion problem could be seen until the width of the tree
was changed.

Update the depth computation such that it is correct every time.

Bug: 132995077
Test: NA
Change-Id: I926af1516769c9451203e7dcda73d457e22259e5