Retry grpc calls that encounter an UNAVAILABLE status
Users of grpc in google3 seem to be retrying this status
exception. It seems to happen especially when the server is
first started but also seem to happen for later requests.

Also add a ping command that sends a small skia image to the
server first. The idea is to make the grpc connection
available with a cheaper call before sending a large skia
image across.

In addition add detection of a stopped skia converter server.
If the server is killed it is detected from the process handle.

Bug: N/A
Test: N//A
Change-Id: Ia593d172863ee25df6567ef46e20f2d063fca1b5