Passing information necessary for scene component
For Compose to be clickable @ component level, we need to recreate
custom scene components. To achieve this, we need the bounds information
within NlModel.

It seems like the easiest and the least registance path to achieve this
is through ViewInfo, which is used by NlModel # ModelUpdater.

This CL intercepts model update, and after the render is complete parses
the necessary information. The custom view info will be later used to
build the scene graph.

Bug: 141027052
Test: Set a break-point and ensure that red-box drawn with debug == on
matches the bounds info parsed.

Change-Id: Ibe7846cd08cff6188295d18eaf799695ad1fb437