Catch NumberFormatException when daemon crashes
When daemon crashed, daemon's standard output contains the
stack trace as in addresses. Studio attempts to send
those backtrace address list as crash info. However,
the last char of the string may be "\n" and will throw
an unhandled NumberFormatException, that will bypass the
daemon's crash-and-restart logic.

This CL doesn't fix b/139791905 because it doesn't stop
daemon crashing but it makes the daemon restart as
expected. In other words, it makes the impact of
some daemon crashes less severe.

Also add logging for simpleperf pre-processing errors.

Bug: 139791905
Test: manual
Change-Id: Ibead384c7be46ec537cd202ea1fff4a3ffa378bf