Display the snapshot in the drop down button
Display the snapshot in the drop down button when a user selects a snapshot
and the SELECT_DEVICE_SNAPSHOT_COMBO_BOX_SNAPSHOTS_ENABLED flag is enabled.
Do not display the snapshot if it's a default one. But do display default
snapshots in drop down submenus as "Quickboot" (as is done in the emulator
UI).

Bug: 120497509
Test: DeviceAndSnapshotComboBoxActionTest
Change-Id: I25fc8fb87471b3fce315e926cb5b30c58346d8a2