Support showing multiple locales in visualization tool
Display all possible locales in the project including the
default locale.

The default locale is used when there
is no particular locale for the device locale.

The project locales includes all found locales in the resource
folders, no matter if the locale folder is layout folder or not.
For example, if there is only 1 layout folder "res/layout"
but there are 2 values folder "res/values/" and "res/values-en",
It shows 2 previews: "Default (no locale)" and "English (en)".

Bug: 141983785
Test: Add LocaleModelsProviderTest
Change-Id: I65744c6428181953f22d016623f41f964344ab1e