Fix GuiPerfLogger hangs at finished()
GuiPerfLogger may hang up at finished() method because it calls
invokeAndWait() without specifying ModalityState. If a test leaves
dialog open (due to failure for instance), invokeAndWait() never
returns. If this happens, the stacktrace of the real test failure
is lost because parent GuiTestRemoteRunner handles this test case
as timed out and the original exception message is not sent to
the parent.

This CL passes ModalityState.any() to invokeAndWait()
method so it does not hang up test execution with dialog.

Bug: 141124739
Fixes: 141761976
Test: none
Change-Id: I363f4527b1b9148296bac09c475e28baaa54a02e