[Dsl] resolving references involves name conversion
Resolving references is done by looking up Dsl elements from external
names.  This necessarily involves converting between the external
representation and our internal model names at the point that we look
up a property in its parent.

This is done at parse time, so get the parser for the current dslFile
and propagate it through to the point where elements are looked up.
Note that we must use the current dslFile's parser, and not the parser
from the parent element's dslFile, because of the possibility of the
property reference being looked up in a language distinct from the
language were the parent element was instantiated.

Also note that this name conversion is a semantic conversion, distinct
from the purely syntactic conversion done by convertReferenceText,
which normalizes external representations of hierarchical properties
to a uniform internal format recognized by GradleNameElement.  For
example, the conversion of

  android.buildTypes.create("xyz").isDebuggable

to

  android.buildTypes.xyz.isDebuggable

is a syntactic transform and is done by convertReferenceText, while

  android.buildTypes.xyz.debuggable

to our canonical model property name

  android.buildTypes.xyz.mDebuggable

can only happen when we know that the xyz is a BuildTypeDslElement

Bug: 130224064
Bug: 141842964
Test: existing (no change in functionality yet; stronger tests to come)
Change-Id: I9e2e097f55faa0d90929a05955316e67e6864463