[Dsl] don't inspect Psi to determine name changes
Instead, keep track of the original name of a property; if the current
name is different from the local name, then the name has been changed
in a way that will require us to write it out.

The original name is also updated after commitNameChanged(), which is
what we want: if we have committed a new name to file, then that new
name is now the effective original name of this property.

Test: existing tests caught broken versions of this
Bug: 141842964
Change-Id: I59200160c6e11fff01cf91e76140997c88e47975