[Dsl] rename all the ProductFlavorModel names
Giving the ProductFlavorModel properties new names involves hooking in
the ProductFlavorDslElement to the external/model name interfaces, by
defining a getExternalToModelMap() method.

There is more machinery here than in the previous renaming in
BuildTypeModel for two reasons:

The first reason is that the ProductFlavor Dsl element supports
handling of several ways of affecting the same conceptual property:
resConfig and resConfigs, testInstrumentationRunnerArgument{,s}.  This
is straightforward to handle with minor modifications because these
methods happen to have the same names and semantics in Kotlin and
Groovy.

The second reason is that some properties have different setters in
Kotlin: there is no dimension `variable'; it can be looked up by value
but cannot be assigned; instead, setDimension() has to be used (and
similarly for testFunctionalTest / testHandleProfiling.  The names in
this renaming are the setter names, which is suitable for setting
properties but not for reading them.  (A bug has been logged to track
fixing this).

Test: existing ProductFlavorModel tests strengthened
Test: some ProductFlavorModel tests implemented for Kotlin
Test: minor modifications to other existing tests
Bug: 130224064
Bug: 141842964
Bug: 142114586
Change-Id: I4b80b4f410486e9166e8fefd21515851448ba6b2