Fix undo not working consistently in the split editor
This CL fixes an issue that was happening when you tried to undo an
operation performed by the new attributes panel. This used to work
before because we would open design files using NlEditor or NavEditor,
and both of them default their implementations of getState(...) to
FileEditor, which in turn returns the same instance everytime. Since the
state in theory doesn't change in this case, we delegate the action of
undoing operations to our own WriteCommandAction mechanism, which does
the job properly.

Now that we introduced the SplitEditor, this stopped working because
SplitEditor inherits from TextEditorWithPreview, which has its own
implementation of getState(...) that returns a different instance
everytime. Performing a CTRL+Z from the attributes panel will cause
UndoRedo to compare the current state with the previous one. Since we
might have two different instances for states in theory equal, we might
consume the undo operation to go back to a state that is essentially the
same we were before.

This particular problem can be fixed by delegating SplitEditor's getState
implementation to its underlying text editor. However, this introduces
another issue: getState() is not only used by IntelliJ to handle the
undo/redo stack, but also to reload the editor's state when you reopen
a file. For split editor, this means restoring the surface mode and the
text editor state (e.g. caret position).

This CL fixes the undo issue and also makes sure to restore the split
editor mode when it opens a file. Restoring the caret position on the
text portion of the split editor will be handled as a separate issue, to
be tackled in a follow up.

Fixes: 140767558
Test: Manual. Select a component in a layout/navigation file. Change
      something in the attributes panel. Ctrl+Z should undo this change.
      Another thing to test is closing/reopen a layout/navigation file
      and observing the mode (e.g. "Split", "Design") is preserved.
Change-Id: I97a400abdc98d1aabf4e7f091c33518e05300f12