[NPW] Add "Use legacy android.support libraries" check-box
Add a new check-box to "Configure Project" Step,
that allows user to use appcompat dependencies.
The checkbox is disabled/unselected for API levels
that don't support appcompat (ie Q) and "androidx only"
Activities (ie Maps Activities).
There is a warning when this check-box is selected,
but "Finish" is still available.

Fixes: 141986023
Test: TemplateTest, existing UI tests and visual inspection of UI
Change-Id: I4173ad82b08232ea52dbb381ebe0ea6c4b190c40