[Dsl] support setConsumerProguardFiles
The support here is not perfect - it is not quite treated equivalently
to an assignment, even though those are the semantics - but it is
enough to allow us to strengthen a Kotlin test and remove a TODO item.

Also make the implementation of consumerProguardFile(s) parallel the
implementation of proguardFile, which respects the appending semantics
of those operators.  Fix the tests that this reveals as broken.

Bug: 72853928
Bug: 141842964
Test: strengthen BuildTypeModel.BuildTypeBlockWithOverrideStatements
Test: fix broken BuildTypeModel and ProductFlavorModel tests
Test: strengthen ProductFlavorModelTest.testSetProguardFiles...
Change-Id: I68db0812e54303c1e0bdb63ec2d28b0724f58b95