Fixes for retry logic
Previously we exited too early because the selected process/stream were
getting set before they were actually connected.
Also we were leaking attach listeners.

Note that exceptions are still seen in the log, but it doesn't seem like
there's a way to prevent that at this point (outside of the transport
framework).

Test: still none
Bug: none
Change-Id: I551a3fb7176abf64e688886d0ae5ea7f072cf07d